feat: initial commit

+++ demo/src/main/java/com/example/minionlinemarket/Repository/ProductRepo.java
@@ -1,9 +1,9 @@
 package com.example.minionlinemarket.Repository;
 
-import com.example.minionlinemarket.model.Product;
+import com.example.minionlinemarket.Model.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProductRepo extends JpaRepository<Product, Integer> {
+public interface ProductRepo extends JpaRepository<Product, Long> {
 }
